Move logger utilities together with the loggers
